Auth for Tasks.

+++ routes/tasks.js
@@ -4,19 +4,19 @@ const express = require('express'),
 const router  = express.Router();
 
 /* GET Tasks listing. */
-router.get('/',async (req, res) =>{
+router.get('/', auth ,async (req, res) =>{
     try {
-        const tasks = await Task.find({});
+        const tasks = await Task.find({owner: req.user._id});
         res.send(tasks);
     } catch (e) {
         res.status(500).send(e);
     }
 });
 
-router.get('/:id', async (req, res, next) =>{
+router.get('/:id', auth, async (req, res) =>{
     const _id = req.params.id;
     try {
-        const task = await Task.findById(_id);
+        const task = await Task.findOne({_id, owner: req.user._id});
         if(!task) {
             return res.status(404).send();
         }
@@ -27,7 +27,7 @@ router.get('/:id', async (req, res, next) =>{
 });
 
 
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowedUpdates = ['description', 'completed'];
     const isValidOperation = updates.every((update) => allowedUpdates.includes(update));
@@ -37,7 +37,7 @@ router.patch('/:id', async (req, res) => {
     }
 
     try {
-        const task = await Task.findById(req.params.id);
+        const task = await Task.findOne({_id: req.params.id, owner: req.user._id});
 
         if (!task) {
             return res.status(404).send()
@@ -67,10 +67,9 @@ router.post('/', auth, async (req, res) => {
     }
 });
 
-router.delete('/:id', async (req, res) => {
-    const _id = req.params.id;
+router.delete('/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findByIdAndDelete(_id);
+        const task = await Task.findByIdAndDelete({_id: req.params.id, owner: req.user._id});
         if(!task) {
             return res.status(404).send();
         }
